tweak example apps to use kotlin api
