+++ compat-kotlin-to-official/src/jvmMain/kotlin/io/embrace/opentelemetry/kotlin/k2j/tracing/TracerProviderAdapter.kt
@@ -4,7 +4,7 @@ import io.embrace.opentelemetry.kotlin.tracing.Tracer
 import io.embrace.opentelemetry.kotlin.tracing.TracerProvider
 import io.opentelemetry.api.OpenTelemetry
 
-internal class TracerProviderAdapter(private val otel: OpenTelemetry) : TracerProvider {
+public class TracerProviderAdapter(private val otel: OpenTelemetry) : TracerProvider {
     override fun getTracer(name: String, version: String?): Tracer {
         val tracerProvider = otel.tracerProvider
         val tracer = when (version) {
