+++ compat-official-to-kotlin/src/jvmTest/kotlin/io/embrace/opentelemetry/kotlin/framework/OtelJavaHarness.kt
@@ -1,0 +1,16 @@
+package io.embrace.opentelemetry.kotlin.framework
+
+import io.embrace.opentelemetry.kotlin.tracing.TracerProvider
+
+internal class OtelJavaHarness {
+
+    val tracerProvider: TracerProvider = TODO("Not implemented yet")
+
+    @Suppress("UNUSED_PARAMETER")
+    internal fun assertSpans(
+        expectedCount: Int,
+        goldenFileName: String? = null
+    ) {
+        TODO("Not implemented yet")
+    }
+}
