+++ examples/android-app-java-api/src/main/AndroidManifest.xml
