+++ examples/android-app-java-api/src/main/kotlin/io/embrace/opentelemetry/example/java/MainActivity.kt
@@ -1,0 +1,5 @@
+package io.embrace.opentelemetry.example.java
+
+import android.app.Activity
+
+class MainActivity : Activity()
