+++ examples/android-app-java-api/src/main/kotlin/io/embrace/opentelemetry/example/java/MainApplication.kt
@@ -1,0 +1,12 @@
+package io.embrace.opentelemetry.example.java
+
+import android.app.Application
+import io.embrace.opentelemetry.example.java.createInstrumentationWithOtelJava
+
+class MainApplication : Application() {
+
+    override fun onCreate() {
+        super.onCreate()
+        createInstrumentationWithOtelJava()
+    }
+}
