+++ examples/android-app-kotlin-api/src/main/AndroidManifest.xml
