+++ examples/android-app-kotlin-api/src/main/kotlin/io/embrace/opentelemetry/example/kotlin/MainActivity.kt
@@ -1,0 +1,5 @@
+package io.embrace.opentelemetry.example.kotlin
+
+import android.app.Activity
+
+class MainActivity : Activity()
