+++ examples/android-app-kotlin-api/src/main/kotlin/io/embrace/opentelemetry/example/kotlin/MainApplication.kt
@@ -1,0 +1,11 @@
+package io.embrace.opentelemetry.example.kotlin
+
+import android.app.Application
+
+class MainApplication : Application() {
+
+    override fun onCreate() {
+        super.onCreate()
+        createInstrumentationWithOtelKotlin()
+    }
+}
