--- examples/android-app/src/main/kotlin/io/embrace/opentelemetry/example/MainActivity.kt
@@ -1,5 +1,0 @@
-package io.embrace.opentelemetry.example
-
-import android.app.Activity
-
-class MainActivity : Activity()
