--- examples/android-app/src/main/kotlin/io/embrace/opentelemetry/example/MainApplication.kt
@@ -1,10 +1,0 @@
-package io.embrace.opentelemetry.example
-
-import android.app.Application
-
-class MainApplication : Application() {
-
-    override fun onCreate() {
-        super.onCreate()
-    }
-}
