--- examples/android-compat-app/src/main/kotlin/io/embrace/opentelemetry/example/compat/MainActivity.kt
@@ -1,5 +1,0 @@
-package io.embrace.opentelemetry.example.compat
-
-import android.app.Activity
-
-class MainActivity : Activity()
