--- examples/android-compat-app/src/main/kotlin/io/embrace/opentelemetry/example/compat/MainApplication.kt
@@ -1,12 +1,0 @@
-package io.embrace.opentelemetry.example.compat
-
-import android.app.Application
-import io.embrace.opentelemetry.example.compat.createInstrumentationWithOtelJava
-
-class MainApplication : Application() {
-
-    override fun onCreate() {
-        super.onCreate()
-        createInstrumentationWithOtelJava()
-    }
-}
