--- examples/app-shared/build.gradle.kts
@@ -1,8 +1,0 @@
-plugins {
-    id("org.jetbrains.kotlin.jvm")
-    id("java")
-}
-
-dependencies {
-    api(project(":opentelemetry-kotlin-api"))
-}
