--- examples/compat-app-shared/build.gradle.kts
@@ -1,11 +1,0 @@
-plugins {
-    id("org.jetbrains.kotlin.jvm")
-    id("java")
-}
-
-dependencies {
-    api(project(":opentelemetry-kotlin-api"))
-    api(platform(libs.opentelemetry.bom))
-    api(libs.opentelemetry.api)
-    implementation(libs.opentelemetry.sdk)
-}
