--- examples/compat-app-shared/src/main/java/io/embrace/opentelemetry/example/compat/ExampleLogRecordProcessor.kt
@@ -1,15 +1,0 @@
-package io.embrace.opentelemetry.example.compat
-
-import io.opentelemetry.context.Context
-import io.opentelemetry.sdk.logs.LogRecordProcessor
-import io.opentelemetry.sdk.logs.ReadWriteLogRecord
-
-internal class ExampleLogRecordProcessor : LogRecordProcessor {
-
-    private val exporter: io.embrace.opentelemetry.example.compat.ExampleLogRecordExporter =
-        io.embrace.opentelemetry.example.compat.ExampleLogRecordExporter()
-
-    override fun onEmit(context: Context, logRecord: ReadWriteLogRecord) {
-        exporter.export(mutableListOf(logRecord.toLogRecordData()))
-    }
-}
