--- examples/compat-app-shared/src/main/java/io/embrace/opentelemetry/example/compat/OtelJavaExample.kt
@@ -1,9 +1,0 @@
-package io.embrace.opentelemetry.example.compat
-
-fun createInstrumentationWithOtelJava() {
-    val api = OtelJavaApi()
-    val span = api.tracer.spanBuilder("my_span").startSpan()
-    span.end()
-    api.logger.logRecordBuilder().setBody("Hello, World!").emit()
-}
-
