+++ examples/example-shared/src/main/java/io/embrace/opentelemetry/example/ExampleLogRecordProcessor.kt
@@ -1,0 +1,15 @@
+package io.embrace.opentelemetry.example
+
+import io.opentelemetry.context.Context
+import io.opentelemetry.sdk.logs.LogRecordProcessor
+import io.opentelemetry.sdk.logs.ReadWriteLogRecord
+
+internal class ExampleLogRecordProcessor : LogRecordProcessor {
+
+    private val exporter: ExampleLogRecordExporter =
+        ExampleLogRecordExporter()
+
+    override fun onEmit(context: Context, logRecord: ReadWriteLogRecord) {
+        exporter.export(mutableListOf(logRecord.toLogRecordData()))
+    }
+}
