+++ examples/example-shared/src/main/java/io/embrace/opentelemetry/example/java/OtelJavaExample.kt
@@ -1,0 +1,9 @@
+package io.embrace.opentelemetry.example.java
+
+fun createInstrumentationWithOtelJava() {
+    val api = OtelJavaApi()
+    val span = api.tracer.spanBuilder("my_span").startSpan()
+    span.end()
+    api.logger.logRecordBuilder().setBody("Hello, World!").emit()
+}
+
