+++ examples/jvm-app-java-api/src/main/java/io/embrace/opentelemetry/example/java/JvmCompatApp.kt
@@ -1,0 +1,5 @@
+package io.embrace.opentelemetry.example.java
+
+fun main() {
+    createInstrumentationWithOtelJava()
+}
