+++ examples/jvm-app-kotlin-api/build.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+    id("org.jetbrains.kotlin.jvm")
+    id("java")
+}
+
+dependencies {
+    implementation(project(":examples:example-shared"))
+}
