+++ examples/jvm-app-kotlin-api/src/main/java/io/embrace/opentelemetry/example/kotlin/JvmApp.kt
@@ -1,0 +1,5 @@
+package io.embrace.opentelemetry.example.kotlin
+
+fun main() {
+    createInstrumentationWithOtelKotlin()
+}
