--- examples/jvm-app/src/main/java/io/embrace/opentelemetry/example/compat/JvmApp.kt
@@ -1,4 +1,0 @@
-package io.embrace.opentelemetry.example.compat
-
-fun main() {
-}
