--- examples/jvm-compat-app/build.gradle.kts
@@ -1,8 +1,0 @@
-plugins {
-    id("org.jetbrains.kotlin.jvm")
-    id("java")
-}
-
-dependencies {
-    implementation(project(":examples:compat-app-shared"))
-}
