--- examples/jvm-compat-app/src/main/java/io/embrace/opentelemetry/example/compat/JvmCompatApp.kt
@@ -1,5 +1,0 @@
-package io.embrace.opentelemetry.example.compat
-
-fun main() {
-    createInstrumentationWithOtelJava()
-}
